requirements.txt
discord.py
PyNaCl
gTTS
yt-dlp
aiofiles